Refactor quality app: move managers from models.py to managers.py
