+++ mrmap/quality/models.py
@@ -8,20 +8,12 @@ Created on: 27.10.20
 from django.db import models, transaction
 from django.urls import reverse
 
-from main.models import CommonInfo, GenericModelMixin
+from main.models import CommonInfo
 from quality.enums import RuleFieldNameEnum, RulePropertyEnum, \
     RuleOperatorEnum, \
     ConformityTypeEnum
+from quality.managers import ConformityCheckRunManager, ConformityCheckConfigurationManager
 from resourceNew.models.metadata import DatasetMetadata
-
-
-class ConformityCheckConfigurationManager(models.Manager):
-    """ Custom manager to extend ConformityCheckConfiguration methods """
-
-    def get_for_metadata_type(self, metadata_type: str):
-        """ Gets all configs that are allowed for the given metadata_type """
-        return super().get_queryset().filter(
-            metadata_types__contains=metadata_type)
 
 
 class ConformityCheckConfiguration(models.Model):
@@ -115,27 +107,6 @@ class ConformityCheckConfigurationInternal(ConformityCheckConfiguration):
     optional_rule_sets = models.ManyToManyField(RuleSet,
                                                 related_name="optional_rule_sets",
                                                 blank=True)
-
-
-class ConformityCheckRunManager(models.Manager):
-    """ Custom manager to extend ConformityCheckRun methods """
-
-    def has_running_check(self, metadata: DatasetMetadata):
-        """ Checks if the given metadata object has a non-finished
-        ConformityCheckRun.
-
-            Returns:
-                True, if a non-finished ConformityCheckRun was found,
-                false otherwise.
-        """
-        running_checks = super().get_queryset().filter(
-            metadata=metadata, passed__isnull=True).count()
-        return running_checks != 0
-
-    def get_latest_check(self, metadata: DatasetMetadata):
-        check = super().get_queryset().filter(metadata=metadata).latest(
-            'time_start')
-        return check
 
 
 class ConformityCheckRun(CommonInfo):
